pyproject.toml
[project]
name = "fastapi-postgres-boilerplate"
version = "0.1.0"
description = "FastAPI boilerplate with PostgreSQL, Redis, and Alembic"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Xolani Mazibuko", email = "mazi76erx@gmail.com" }]
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.42.0",
    "sqlalchemy>=2.0.48",
    "alembic>=1.18.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "redis>=7.3.0",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.11",
    "fastapi-debug-toolbar>=0.6.3",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
    "types-redis>=4.6.0.20241004",
    "pre-commit>=4.5.1",
]

[project.scripts]
start = "server.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.ruff]
target-version = "py314"
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "ARG"]
ignore = ["E501", "B008", "C901", "ARG001"]

[tool.ruff.lint.isort]
known-first-party = ["server"]

[tool.ruff.lint.per-file-ignores]
"server/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.14"
files = ["server"]
ignore_missing_imports = true
allow_redefinition = true
disallow_untyped_decorators = false
explicit_package_bases = true
warn_unused_ignores = true
check_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = ["server"]
addopts = ["--strict-markers", "-ra"]
markers = ["integration: mark test as integration test"]

[tool.coverage.run]
branch = true
source = ["server"]

[tool.coverage.report]
ignore_errors = true

[tool.coverage.html]
directory = "htmlcov"